add move semantics to mutexes
This commit changes a lot of things about mutexes, they work the same,
however now they allow for moving mutexes. This means the underlying
mutex now lives in a unique_ptr. This comes at a (relatively) tiny
runtime cost, but the tradeoff is worth it in most scenarios
(specifically) those which have large(r) amounts of data in the mutex.
A few corners have been cut (on purpose) for allowing minimal cost while
keeping safety 100% (hopefully.)